+++ Examples/apps/ObjCExample/ObjCExample/AppDelegate.m
@@ -8,6 +8,8 @@
 #import "AppDelegate.h"
 
 @import Segment;
+@import SegmentObjCSupport;
+@import SegmentMixpanel;
 
 @interface AppDelegate ()
 
@@ -39,6 +41,9 @@
         }
         return event;
     }];
+    
+    [self.analytics addDestination:[[SEGMixpanelDestination alloc] init]];
+    
     
     dispatch_async(dispatch_get_main_queue(), ^{
         [self.analytics track:@"booya"];
