+++ Sources/Segment/ObjC/ObjCAnalytics.swift
@@ -79,7 +79,7 @@ extension ObjCAnalytics {
     
     /// Track a screen change with a title, category and other properties.
     /// - Parameters:
-    ///   - screenTitle: The title of the screen being tracked.
+    ///   - title: The title of the screen being tracked.
     @objc(screen:)
     public func screen(title: String) {
         screen(title: title, category: nil, properties: nil)
@@ -87,7 +87,7 @@ extension ObjCAnalytics {
     
     /// Track a screen change with a title, category and other properties.
     /// - Parameters:
-    ///   - screenTitle: The title of the screen being tracked.
+    ///   - title: The title of the screen being tracked.
     ///   - category: A category to the type of screen if it applies.
     @objc(screen:category:)
     public func screen(title: String, category: String?) {
@@ -95,7 +95,7 @@ extension ObjCAnalytics {
     }
     /// Track a screen change with a title, category and other properties.
     /// - Parameters:
-    ///   - screenTitle: The title of the screen being tracked.
+    ///   - title: The title of the screen being tracked.
     ///   - category: A category to the type of screen if it applies.
     ///   - properties: Any extra metadata associated with the screen. e.g. method of access, size, etc.
     @objc(screen:category:properties:)
@@ -129,62 +129,6 @@ extension ObjCAnalytics {
     }
 }
 
-// MARK: - ObjC Plugin Functionality
-
-@objc
-extension ObjCAnalytics {
-    /// This method allows you to add middleware to an Analytics instance, similar to Analytics-iOS.
-    /// However, it is **strongly encouraged** that Enrichments/Plugins/Middlewares be written in swift
-    /// to avoid the overhead of type conversion back and forth.  This exists solely for compatibility
-    /// purposes.
-    ///
-    /// Example:
-    ///    [self.analytics addSourceMiddleware:^NSDictionary<NSString *,id> * _Nullable(NSDictionary<NSString *,id> * _Nullable event) {
-    ///        // drop all events named booya
-    ///        NSString *eventType = event[@"type"];
-    ///        if ([eventType isEqualToString:@"track"]) {
-    ///            NSString *eventName = event[@"event"];
-    ///            if ([eventName isEqualToString:@"booya"]) {
-    ///                return nil;
-    ///            }
-    ///        }
-    ///        return event;
-    ///    }];
-    ///
-    /// - Parameter middleware: The middleware to execute at the source level.
-    @objc(addSourceMiddleware:)
-    public func addSourceMiddleware(middleware: @escaping ((_ event: [String: Any]?) -> [String: Any]?)) {
-        analytics.add(plugin: ObjCShimPlugin(middleware: middleware))
-    }
-    
-    /// This method allows you to add middleware to an Analytics instance, similar to Analytics-iOS.
-    /// However, it is **strongly encouraged** that Enrichments/Plugins/Middlewares be written in swift
-    /// to avoid the overhead of type conversion back and forth.  This exists solely for compatibility
-    /// purposes.
-    ///
-    /// Example:
-    ///    [self.analytics addDestinationMiddleware:^NSDictionary<NSString *,id> * _Nullable(NSDictionary<NSString *,id> * _Nullable event) {
-    ///        // drop all events named booya on the amplitude destination
-    ///        NSString *eventType = event[@"type"];
-    ///        if ([eventType isEqualToString:@"track"]) {
-    ///            NSString *eventName = event[@"event"];
-    ///            if ([eventName isEqualToString:@"booya"]) {
-    ///                return nil;
-    ///            }
-    ///        }
-    ///        return event;
-    ///    }, forKey: @"Amplitude"];
-    ///
-    /// - Parameters:
-    ///   - middleware: The middleware to execute at the source level.
-    ///   - destinationKey: A string value representing the destination.  ie: @"Amplitude"
-    @objc(addDestinationMiddleware:forKey:)
-    public func addDestinationMiddleware(middleware: @escaping ((_ event: [String: Any]?) -> [String: Any]?), destinationKey: String) {
-        // couldn't find the destination they wanted
-        guard let dest = analytics.find(key: destinationKey) else { return }
-        _ = dest.add(plugin: ObjCShimPlugin(middleware: middleware))
-    }
-}
 
 // MARK: - ObjC Peripheral Functionality
